[Linux] Make complain support kdialog
Note that the version of kdialog in debian-live-6.0.1-i386-kde-desktop.iso
does not set the window size properly for five or more lines of text.
We may or may not run into that bug in this script, but there's nothing
we can do about it (except possibly report it upstream).