W32: Move patch-mozbuild to install target
also prevent double-patching the source. Doing the latter makes sense,
because otherwise a patch might get silently reversed. Doing the former
makes sense because after extraction of mozillabuild, we need to patch
it before we can use it further; and when we change the patch we need to
re-extract to prevent double-patching

While we do that, also set up prerequisites so mozilla-build gets remade
when the patches change