Put empty targets in an arch-dependent location
Otherwise make is confused when ARCH_TYPE is changed with existing empty
targets. STAMP_DIR is added for clarity but no target is added as it would
probably create a cycle with FETCH_DIR. It could easily be added afterwards.
vpath directives are picked over VPATH because there are many build products
and it is better to risk rebuilding than selecting targets by mistake.