Apply the openssl OSX patch in unpack target
This means we now also apply the patch for linux and windows, but that
should be fine - it doesn't touch files they need. This means for us
that if the build fails and we have to re-run the build-openssl target
with a changed configuration or so, we can without re-extracting.