Make TBB work on KDE 4
On KDE 4 desktops, Qt apps try to load the KDE 4 style called Oxygen.
This causes a few other Qt shared libs to be loaded, and we don't ship
them all, which causes a crash. Tomásfound this cool solution, which
means we don't have to ship more libraries than we do now. Basically we
tell Vidalia to bypass auto-detecting a style and just pre-pick a sane
one. If the user manually overrides this, Vidalia will still crash, but
this should work for the vast majority of users for now.

Works around bug 5214.